+++ modules/auth/sso/sso.go
@@ -6,11 +6,13 @@
 package sso
 
 import (
+	"fmt"
 	"reflect"
 	"strings"
 
 	"code.gitea.io/gitea/models"
 	"code.gitea.io/gitea/modules/log"
+	"code.gitea.io/gitea/modules/setting"
 
 	"gitea.com/macaron/macaron"
 	"gitea.com/macaron/session"
@@ -18,6 +20,14 @@ import (
 
 var (
 	ssoMethods []SingleSignOn
+)
+
+// The purpose of the following three function variables is to let the linter know that
+// those functions are not dead code and are actually being used
+var (
+	_ = isPublicResource
+	_ = isPublicPage
+	_ = handleSignIn
 )
 
 // Methods returns the instances of all registered SSO methods
@@ -83,6 +93,79 @@ func isAPIPath(ctx *macaron.Context) bool {
 // isAttachmentDownload check if request is a file download (GET) with URL to an attachment
 func isAttachmentDownload(ctx *macaron.Context) bool {
 	return strings.HasPrefix(ctx.Req.URL.Path, "/attachments/") && ctx.Req.Method == "GET"
+}
+
+// isPublicResource checks if the url is of a public resource file that should be served
+// without authentication (eg. the Web App Manifest, the Service Worker script or the favicon)
+func isPublicResource(ctx *macaron.Context) bool {
+	path := strings.TrimSuffix(ctx.Req.URL.Path, "/")
+	return path == "/robots.txt" ||
+		path == "/favicon.ico" ||
+		path == "/favicon.png" ||
+		path == "/manifest.json" ||
+		path == "/serviceworker.js"
+}
+
+// isPublicPage checks if the url is of a public page that should not require authentication
+func isPublicPage(ctx *macaron.Context) bool {
+	path := strings.TrimSuffix(ctx.Req.URL.Path, "/")
+	homePage := strings.TrimSuffix(setting.AppSubURL, "/")
+	currentURL := homePage + path
+	return currentURL == homePage ||
+		path == "/user/login" ||
+		path == "/user/login/openid" ||
+		path == "/user/sign_up" ||
+		path == "/user/forgot_password" ||
+		path == "/user/openid/connect" ||
+		path == "/user/openid/register" ||
+		strings.HasPrefix(path, "/user/oauth2") ||
+		path == "/user/link_account" ||
+		path == "/user/link_account_signin" ||
+		path == "/user/link_account_signup" ||
+		path == "/user/two_factor" ||
+		path == "/user/two_factor/scratch" ||
+		path == "/user/u2f" ||
+		path == "/user/u2f/challenge" ||
+		path == "/user/u2f/sign" ||
+		(!setting.Service.RequireSignInView && (path == "/explore/repos" ||
+			path == "/explore/users" ||
+			path == "/explore/organizations" ||
+			path == "/explore/code"))
+}
+
+// handleSignIn clears existing session variables and stores new ones for the specified user object
+func handleSignIn(ctx *macaron.Context, sess session.Store, user *models.User) {
+	_ = sess.Delete("openid_verified_uri")
+	_ = sess.Delete("openid_signin_remember")
+	_ = sess.Delete("openid_determined_email")
+	_ = sess.Delete("openid_determined_username")
+	_ = sess.Delete("twofaUid")
+	_ = sess.Delete("twofaRemember")
+	_ = sess.Delete("u2fChallenge")
+	_ = sess.Delete("linkAccount")
+	err := sess.Set("uid", user.ID)
+	if err != nil {
+		log.Error(fmt.Sprintf("Error setting session: %v", err))
+	}
+	err = sess.Set("uname", user.Name)
+	if err != nil {
+		log.Error(fmt.Sprintf("Error setting session: %v", err))
+	}
+
+	// Language setting of the user overwrites the one previously set
+	// If the user does not have a locale set, we save the current one.
+	if len(user.Language) == 0 {
+		user.Language = ctx.Locale.Language()
+		if err := models.UpdateUserCols(user, "language"); err != nil {
+			log.Error(fmt.Sprintf("Error updating user language [user: %d, locale: %s]", user.ID, user.Language))
+			return
+		}
+	}
+
+	ctx.SetCookie("lang", user.Language, nil, setting.AppSubURL, setting.SessionConfig.Domain, setting.SessionConfig.Secure, true)
+
+	// Clear whatever CSRF has right now, force to generate a new one
+	ctx.SetCookie(setting.CSRFCookieName, "", -1, setting.AppSubURL, setting.SessionConfig.Domain, setting.SessionConfig.Secure, true)
 }
 
 // init populates the list of SSO authentication plugins in the order they are expected to be
